Fix unicode handling, unclosed block comments, and error messages
oss-fuzz found an error in how some of the code processed unicode
characters. Initially, this was fixed by replacing
`str.chars().enumerate()` with the more correct `str.char_indices()`.

But after fixing that issue, the selected fuzz test case also
highlighted a parser issue when handling unclosed block comments. This
is an error, but should have been reported as a parsing error (an error
in the input document). Instead it was reported as an internal parser
error.

After correcting the error, I also uncovered an issue in the code for
computing the next line to be parsed (for referencing the source
location of syntax errors). Coincidentally, the fix was in the same
function that had the unicode issue, and fixing it removed the need for
`char_indices()`.

Tests were added to validate the unicode fixes and the unclosed comment
block error handling.