Limit nesting to prevent program stack overflow, and catch bad brace
This PR addresses two more classes of oss-fuzz failures.

1. A document with many open braces, without closing braces, can
eventually cause program stack overflow. I've added function variations
to set a specific limit, and the existing functions set this value to a
default limit.

2. A document that ends with an extra closing brace could crash the
parser. This is now caught, and a parser error is emitted. This may have
also contributed to an "out-of-memory" error, as the reproducer test
(included with this PR) no longer fails.

I also added better testing support for debugging the parser on sample
documents (such as oss-fuzz reproducers).

Fixes: #22
Fixes: #23
Fixes: #24